changed file structure of the public and private views
