+++ src/views/public/LandingPage/Home.jsx
@@ -1,4 +1,4 @@
-import Shape from '../components/Shape';
+import Shape from '../../../components/Shape';
 
 export default function Home() {
   return (
